scraper/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-jobspy==1.1.75
